add django tasks
