+++ webapp/Procfile
@@ -1,2 +1,3 @@
 web: bash bin/start
+worker: python manage.py db_worker
 postdeploy: bash bin/post_deploy
