+++ webapp/pyproject.toml
@@ -9,6 +9,7 @@ dependencies = [
     "dj-database-url>=3.0.1,<4",
     "django>=6.0.6",
     "django-extensions~=4.1",
+    "django-tasks>=0.9.0,<1",
     "django-import-export[xls, xlsx]>=4.4.1,<5",
     "django-ninja>=1.6.2,<2",
     "django-storages>=1.14.6,<2",
